// src/webflow/code-components/product-colors.types.ts
import type { WebflowImageAsset } from "./webflow-asset.types";

export type ProductColorsWebflowProps = {
  ariaLabel?: string;
  color1Alt?: string;
  color1Image?: WebflowImageAsset;
  color1Label?: string;
  color1Swatch?: WebflowImageAsset;
  color2Alt?: string;
  color2Image?: WebflowImageAsset;
  color2Label?: string;
  color2Swatch?: WebflowImageAsset;
  color3Alt?: string;
  color3Image?: WebflowImageAsset;
  color3Label?: string;
  color3Swatch?: WebflowImageAsset;
  color4Alt?: string;
  color4Image?: WebflowImageAsset;
  color4Label?: string;
  color4Swatch?: WebflowImageAsset;
  visibleColorCount?: number;
};